did the routes, have some commented questions

+++ Develop/routes/api/category-routes.js
@@ -3,35 +3,54 @@ const { Category, Product } = require('../../models');
 
 // The `/api/categories` endpoint
 
-router.get('/', (req, res) => {
+router.get('/', async (req, res) => {
   // find all categories
   // be sure to include its associated Products
-  Category.findAll().then((catInfo) => {
-    res.json(catInfo);
-  });
+  try {
+    const categoryData = await Category.findAll({
+      include: [{ model: Category, model: Product}],
+    });
+    res.status(200).json(categoryData);
+  } catch (err) {
+    res.status(500).json(err);
+  }
+  
 });
 // refer to instructional #7 for the parameter in the .then above^^
-router.get('/:id', (req, res) => {
+router.get('/:id', async (req, res) => {
   // find one category by its `id` value
   // be sure to include its associated Products
-});
-
-router.post('/', (req, res) => {
-  // create a new category
-  //working on 
-  Category.create({
-    id: req.body.id,
-    name: req.body.name
-
-  }).then((newCategory) => {
-    req.json(newCategory);
-  })
-
-    .catch((err) => {
-      res.json(err);
+  try {
+    const categoryData = await Category.findByPk(req.params.id, {
+      include: [{ model: Category, model: Product}],
+      // might not need the Category model above^^
     });
+    if (!categoryData) {
+      res.status(404).json({ message: 'Not category with that id!'});
+      return;
+    }
+    res.status(200).json(categoryData);
+  } catch (err) {
+    res.status(500).json(err);
+  }
 });
-// reference from class
+
+// router.post('/', async (req, res) => {
+//   // create a new category
+//   //working on 
+// //   Category.create({
+// //     id: req.body.id,
+// //     name: req.body.name
+
+// //   }).then((newCategory) => {
+// //     req.json(newCategory);
+// //   })
+
+// //     .catch((err) => {
+// //       res.json(err);
+// //     });
+// // });
+// // reference from class
 router.post('/', async (req, res) => {
   try {
     const catCreate = await Category.create(req.body);
@@ -43,32 +62,41 @@ router.post('/', async (req, res) => {
   }
 });
 
-router.put('/:id', (req, res) => {
+router.put('/:id', async (req, res) => {
   // update a category by its `id` value
-  Category.update({
-    catagory_id: id.body.id
-  },
-    {
-      where: {
-        catagory_id: req.params.id,
-        name: req.params.name
-      },
-
-    })
+  try {
+ const categoryData = await Category.update(req.body ,{
+  where: {
+    catagory_id: req.params.id,
+ },
+});
+if (categoryData[0]) {
+  res.status(404).json({ message: 'No user with this id!'});
+  return;
+}
+res.status(200).json(categoryData);
+  } catch (err) {
+    res.status(500).json(err);
+  }
 });
 
-router.delete('/:id', (req, res) => {
+router.delete('/:id', async (req, res) => {
   // delete a category by its `id` value
-  Category.destroy({
+  try {
+  const categoryData = await Category.destroy({
     where: {
-      id: req.body.id,
-    }
-  })
-    .then((deletedCategory) => {
-      res.json(deletedCategory);
-    })
-    .catch((err) => res.json(err))
-
+      catagory_id: req.body.id,
+      // id :req.params.id, ?? which one do i need catagory_id or id?
+    },
+  });
+    if (!deletedCategory) {
+      res.status(404).json({ message: 'No reader found with that id!' });
+      return;
+  }
+  res.status(200).json(categoryData);
+} catch (err) {
+  res.status(500).json(err);
+}
 });
 
 module.exports = router;
